pyproject.toml
[tool.poetry]
name = "aton-test"
version = "0.1.0"
description = ""
authors = ["OlegKhiretdinov <oleg@oleg.win>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
faker = "^25.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
